+++ app/commands/base.rb
