+++ app/commands/speakers.rb
