+++ app/commands/vote.rb
