+++ lib/repositories/talk_repo.rb
@@ -1,0 +1,9 @@
+module Repositories
+  class TalkRepo < ROM::Repository[:talks]
+    include ArgsImport['rom']
+
+    def later_that_day
+      talks.by_datetime(DateTime.now)
+    end
+  end
+end
