+++ system/boot/rom.rb
@@ -1,16 +1,13 @@
-require 'relations/jobs'
-require 'relations/talks'
-
-Application.boot(:rom) do
+Application.boot(:rom) do |app|
   init do
     require 'rom'
+    require 'rom-repository'
   end
 
   start do
-    configuration = ROM::Configuration.new(:yaml, 'data')
-    configuration.register_relation(Relations::Jobs)
-    configuration.register_relation(Relations::Talks)
-    # configuration.auto_registration('lib')
+    configuration = ROM::Configuration.new(:yaml, app.root.join('data'))
+    configuration.auto_registration(app.root.join('lib/persistence'), namespace: false)
+
     container = ROM.container(configuration)
     register(:rom, container)
   end
