--- system/import.rb
@@ -1,2 +1,0 @@
-require_relative 'container'
-Import = Application.injector
